% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_colorio}
\alias{check_colorio}
\title{Check the colorio installation}
\usage{
check_colorio(quiet = FALSE)
}
\arguments{
\item{quiet}{\code{logical}, if \code{TRUE}, suppresses message upon successful check}
}
\value{
invisible \code{NULL}, called for side-effects
}
\description{
To install colorio into a Python environment, use \code{\link[=install_colorio]{install_colorio()}}.
}
\examples{
if (FALSE) {
  # not run because it requires Python
  check_colorio()
}
}
\seealso{
\code{\link[reticulate:py_config]{reticulate::py_config()}}, \code{\link[=install_colorio]{install_colorio()}}, \code{\link[=colorio_version]{colorio_version()}}
}
