% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorio-package.R
\name{colorio-package}
\alias{colorio-package}
\title{colorio: package to wrap colorio Python package}
\description{
This package offers low-level access to the
\href{https://github.com/nschloe/colorio}{colorio} Python package, using the
\href{https://rstudio.github.io/reticulate}{reticulate} package.
}
\details{
The immediate motivation is to provide R users wuth access to the wide
array of color spaces available in colorio, including many newer ones like
CIECAM02, CAM16, and Jzazbz.
}
