% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{colorio}
\alias{colorio}
\title{colorio object}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 5.
}
\usage{
colorio
}
\description{
Uses the reticulate framework to access the colorio API.
}
\details{
The color Python package is exposed through the \code{colorio} object.

In this package, use the \code{$} operator wherever you see the \code{.} operator
used in Python.
}
\keyword{datasets}
