% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{colorio_version}
\alias{colorio_version}
\title{Installed version of colorio}
\usage{
colorio_version()
}
\value{
\code{character} version
}
\description{
Installed version of colorio
}
\examples{
if (interactive()) {
  colorio_version()
}

}
