% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_colorio}
\alias{install_colorio}
\title{Install colorio Python package}
\usage{
install_colorio(
  method = c("conda", "virtualenv"),
  envname = "r-reticulate",
  ...
)
}
\arguments{
\item{method}{\code{character}, indicates to use \code{"conda"} or \code{"virtualenv"}}

\item{envname}{\code{character}, name of environment into which to install}

\item{...}{other arguments sent to \code{\link[reticulate:conda-tools]{reticulate::conda_install()}} or
\code{\link[reticulate:virtualenv-tools]{reticulate::virtualenv_install()}}}
}
\value{
Invisible \code{NULL}, called for side effects
}
\description{
This function wraps installation functions from \link[reticulate:reticulate]{reticulate} to install the Python package
\strong{colorio}.
}
\details{
This package uses the \link[reticulate:reticulate]{reticulate} package
to make an interface with the \href{https://github.com/nschloe/colorio}{colorio}
Python package. To promote consistency in usage of \strong{reticulate} among
different R packages, it is
\href{https://rstudio.github.io/reticulate/articles/package.html#installing-python-dependencies}{recommended}
to use a common Python environment, called \code{"r-reticulate"}.

Depending on your setup, you can create this environment using
\code{\link[reticulate:conda-tools]{reticulate::conda_create()}} or \code{\link[reticulate:virtualenv-tools]{reticulate::virtualenv_create()}},
as described in this
\href{https://rstudio.github.io/reticulate/articles/python_packages.html#conda-installation}{reticulate article},
or in this package's \href{https://vegawidget.github.io/altair/articles/installation.html#python-env}{Installation article}.
}
\examples{
if (FALSE) {
  # not run because it requires Python
  install_colorio()
}
}
